Streaming/requirements.txt
pandas
streamlit
wordcloud
matplotlib
textblob
python-dotenv
praw
boto3
altair